refactor(gui): Move shared models to app/models
