+++ lib/gui/app/app.js
@@ -36,13 +36,13 @@ const release = require('../../shared/release')
 const store = require('./models/store')
 const errors = require('../../shared/errors')
 const packageJSON = require('../../../package.json')
-const flashState = require('../../shared/models/flash-state')
+const flashState = require('./models/flash-state')
 const settings = require('./models/settings')
 const windowProgress = require('./os/window-progress')
 const analytics = require('./modules/analytics')
 const updateNotifier = require('./components/update-notifier')
-const availableDrives = require('../../shared/models/available-drives')
-const selectionState = require('../../shared/models/selection-state')
+const availableDrives = require('./models/available-drives')
+const selectionState = require('./models/selection-state')
 const driveScanner = require('./modules/drive-scanner')
 const osDialog = require('./os/dialog')
 const exceptionReporter = require('./modules/exception-reporter')
