+++ lib/gui/app/components/file-selector/file-selector/file-selector.jsx
@@ -37,7 +37,7 @@ const Storage = require('../../../models/storage')
 const analytics = require('../../../modules/analytics')
 const middleEllipsis = require('../../../utils/middle-ellipsis')
 const files = require('../../../../../shared/files')
-const selectionState = require('../../../../../shared/models/selection-state')
+const selectionState = require('../../../models/selection-state')
 const imageStream = require('../../../../../sdk/image-stream')
 const errors = require('../../../../../shared/errors')
 const messages = require('../../../../../shared/messages')
