+++ lib/gui/app/components/flash-error-modal/services/flash-error-modal.js
@@ -16,8 +16,8 @@
 
 'use strict'
 
-const flashState = require('../../../../../shared/models/flash-state')
-const selectionState = require('../../../../../shared/models/selection-state')
+const flashState = require('../../../models/flash-state')
+const selectionState = require('../../../models/selection-state')
 const analytics = require('../../../modules/analytics')
 
 module.exports = function (WarningModalService) {
