+++ lib/gui/app/models/available-drives.js
@@ -17,7 +17,7 @@
 'use strict'
 
 const _ = require('lodash')
-const store = require('../../gui/app/models/store')
+const store = require('./store')
 
 /**
  * @summary Check if there are available drives
