+++ lib/gui/app/models/flash-state.js
@@ -17,8 +17,8 @@
 'use strict'
 
 const _ = require('lodash')
-const store = require('../../gui/app/models/store')
-const units = require('../units')
+const store = require('./store')
+const units = require('../../../shared/units')
 
 /**
  * @summary Reset flash state
