+++ lib/gui/app/models/selection-state.js
@@ -17,7 +17,7 @@
 'use strict'
 
 const _ = require('lodash')
-const store = require('../../gui/app/models/store')
+const store = require('./store')
 const availableDrives = require('./available-drives')
 
 /**
