+++ lib/gui/app/modules/image-writer.js
@@ -24,14 +24,14 @@ const ipc = require('node-ipc')
 const isRunningInAsar = require('electron-is-running-in-asar')
 const electron = require('electron')
 const settings = require('../models/settings')
-const flashState = require('../../../shared/models/flash-state')
+const flashState = require('../models/flash-state')
 const errors = require('../../../shared/errors')
 const permissions = require('../../../shared/permissions')
 const windowProgress = require('../os/window-progress')
 const analytics = require('../modules/analytics')
 const updateLock = require('./update-lock')
 const packageJSON = require('../../../../package.json')
-const selectionState = require('../../../shared/models/selection-state')
+const selectionState = require('../models/selection-state')
 
 /**
  * @summary Number of threads per CPU to allocate to the UV_THREADPOOL
