+++ lib/gui/app/pages/finish/controllers/finish.js
@@ -18,8 +18,8 @@
 
 const _ = require('lodash')
 const settings = require('../../../models/settings')
-const flashState = require('../../../../../shared/models/flash-state')
-const selectionState = require('../../../../../shared/models/selection-state')
+const flashState = require('../../../models/flash-state')
+const selectionState = require('../../../models/selection-state')
 const analytics = require('../../../modules/analytics')
 const updateLock = require('../../../modules/update-lock')
 const messages = require('../../../../../shared/messages')
