+++ lib/gui/app/pages/main/controllers/drive-selection.js
@@ -20,7 +20,7 @@ const _ = require('lodash')
 const angular = require('angular')
 const prettyBytes = require('pretty-bytes')
 const settings = require('../../../models/settings')
-const selectionState = require('../../../../../shared/models/selection-state')
+const selectionState = require('../../../models/selection-state')
 const analytics = require('../../../modules/analytics')
 const exceptionReporter = require('../../../modules/exception-reporter')
 const utils = require('../../../../../shared/utils')
