+++ lib/gui/app/pages/main/controllers/flash.js
@@ -18,7 +18,7 @@
 
 const _ = require('lodash')
 const messages = require('../../../../../shared/messages')
-const flashState = require('../../../../../shared/models/flash-state')
+const flashState = require('../../../models/flash-state')
 const driveScanner = require('../../../modules/drive-scanner')
 const progressStatus = require('../../../modules/progress-status')
 const notification = require('../../../os/notification')
@@ -27,7 +27,8 @@ const imageWriter = require('../../../modules/image-writer')
 const path = require('path')
 const store = require('../../../models/store')
 const constraints = require('../../../../../shared/drive-constraints')
-const availableDrives = require('../../../../../shared/models/available-drives')
+const availableDrives = require('../../../models/available-drives')
+const debug = require('debug')('etcher:controller:flash')
 
 module.exports = function (
   $q,
@@ -139,7 +140,10 @@ module.exports = function (
 
     // Trigger Angular digests along with store updates, as the flash state
     // updates. Without this there is essentially no progress to watch.
-    const unsubscribe = store.subscribe($timeout)
+    const unsubscribe = store.subscribe(() => {
+      debug('store.onChange')
+      $timeout()
+    })
 
     const iconPath = '../../../assets/icon.png'
