+++ lib/gui/app/pages/main/controllers/image-selection.js
@@ -24,7 +24,7 @@ const errors = require('../../../../../shared/errors')
 const imageStream = require('../../../../../sdk/image-stream')
 const supportedFormats = require('../../../../../shared/supported-formats')
 const analytics = require('../../../modules/analytics')
-const selectionState = require('../../../../../shared/models/selection-state')
+const selectionState = require('../../../models/selection-state')
 const osDialog = require('../../../os/dialog')
 const exceptionReporter = require('../../../modules/exception-reporter')
