+++ lib/gui/app/pages/main/controllers/main.js
@@ -17,11 +17,11 @@
 'use strict'
 
 const settings = require('../../../models/settings')
-const flashState = require('../../../../../shared/models/flash-state')
+const flashState = require('../../../models/flash-state')
 const analytics = require('../../../modules/analytics')
 const exceptionReporter = require('../../../modules/exception-reporter')
-const availableDrives = require('../../../../../shared/models/available-drives')
-const selectionState = require('../../../../../shared/models/selection-state')
+const availableDrives = require('../../../models/available-drives')
+const selectionState = require('../../../models/selection-state')
 const driveConstraints = require('../../../../../shared/drive-constraints')
 const messages = require('../../../../../shared/messages')
