+++ tests/shared/models/available-drives.spec.js
@@ -18,8 +18,8 @@
 
 const m = require('mochainon')
 const path = require('path')
-const availableDrives = require('../../../lib/shared/models/available-drives')
-const selectionState = require('../../../lib/shared/models/selection-state')
+const availableDrives = require('../../../lib/gui/app/available-drives')
+const selectionState = require('../../../lib/gui/app/selection-state')
 const constraints = require('../../../lib/shared/drive-constraints')
 
 describe('Model: availableDrives', function () {
